EGATE-83: Event Adapter Deployment
- Renamed activiti-cloud-events-adapter-service to
activiti-cloud-services-events-adapter to follow Activiti naming
precedence
- Changed packages to org.activiti.cloud.services to follow Activiti
naming precedence
- Changed Activiti version to 7.0.0-SNAPSHOT
- Updated events to reference 7.0.0-SNAPSHOT changes
- Added EnableActivitiEventAdapter class
