--- activiti-cloud-events-adapter-service/src/main/java/org/activiti/cloud/events/adapter/EventGateway.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.activiti.cloud.events.adapter;
-
-import org.alfresco.event.model.EventV1;
-import org.alfresco.event.model.ResourceV1;
-import org.springframework.integration.annotation.Gateway;
-import org.springframework.integration.annotation.MessagingGateway;
-
-@MessagingGateway
-public interface EventGateway {
-    String EVENTS_ADAPTER_PRODUCER_CHANNEL = "eventsAdapterProducerChannel";
-
-    @Gateway(requestChannel = EVENTS_ADAPTER_PRODUCER_CHANNEL)
-    void send(EventV1<ResourceV1> event);
-}
