--- activiti-cloud-events-adapter-service/src/main/java/org/activiti/cloud/events/adapter/streams/EventsAdapterConsumerChannel.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.activiti.cloud.events.adapter.streams;
-
-import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.messaging.SubscribableChannel;
-
-public interface EventsAdapterConsumerChannel {
-
-    String EVENTS_ADAPTER_CONSUMER = "eventsAdapterConsumer";
-
-    @Input(EVENTS_ADAPTER_CONSUMER)
-    SubscribableChannel eventsAdapterConsumer();
-}
