--- activiti-cloud-events-adapter-service/src/main/java/org/activiti/cloud/events/adapter/transformers/GenericEventTransformer.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.activiti.cloud.events.adapter.transformers;
-
-import org.activiti.runtime.api.event.CloudRuntimeEvent;
-import org.alfresco.event.model.activiti.ActivitiCloudRuntimeResourceV1;
-import org.alfresco.event.model.EventV1;
-import org.springframework.stereotype.Component;
-
-@Component
-public class GenericEventTransformer extends AbstractEventTransformer<CloudRuntimeEvent, ActivitiCloudRuntimeResourceV1> {
-
-    public static final String UNKNOWN_TYPE = "UNKNOWN_TYPE";
-
-    public GenericEventTransformer() {
-        super(UNKNOWN_TYPE);
-    }
-
-    @Override
-    public EventV1<ActivitiCloudRuntimeResourceV1> transform(CloudRuntimeEvent event) {
-        ActivitiCloudRuntimeResourceV1 resource = new ActivitiCloudRuntimeResourceV1(event.getId(), null);
-        setCommonValues(event, resource);
-
-        return new EventV1<>(event.getEventType().name(), null, resource);
-    }
-}
