+++ activiti-cloud-services-events-adapter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  org.activiti.cloud.services.events.adapter.jms.JmsConfig
