+++ activiti-cloud-starter-events-adapter/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>activiti-cloud-events-adapter-service-parent</artifactId>
         <groupId>org.activiti.cloud</groupId>
-        <version>7.0.0.Beta1</version>
+        <version>7.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -13,19 +13,22 @@
     <name>Activiti Cloud Events Adapter :: Starter</name>
 
     <properties>
-        <activiti-cloud-events-adapter-service.version>0.0.1-SNAPSHOT</activiti-cloud-events-adapter-service.version>
         <dependency.activemq-junit.version>5.15.5</dependency.activemq-junit.version>
     </properties>
 
     <dependencies>
         <dependency>
+            <groupId>org.activiti.cloud</groupId>
+            <artifactId>activiti-cloud-services-events-adapter</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.activiti.cloud</groupId>
-            <artifactId>activiti-cloud-events-adapter-service</artifactId>
-            <version>${activiti-cloud-events-adapter-service.version}</version>
+          <groupId>org.springframework.cloud</groupId>
+          <artifactId>spring-cloud-commons</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
