+++ pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.activiti.cloud</groupId>
         <artifactId>activiti-cloud-parent</artifactId>
-        <version>7.0.0.Beta1</version>
+        <version>7.0.0-SNAPSHOT</version>
         <relativePath/>
     </parent>
     <artifactId>activiti-cloud-events-adapter-service-parent</artifactId>
@@ -13,7 +13,7 @@
     <name>Activiti Cloud Events Adapter :: Parent</name>
     <url>http://activiti.org</url>
     <modules>
-        <module>activiti-cloud-events-adapter-service</module>
+        <module>activiti-cloud-services-events-adapter</module>
         <module>activiti-cloud-starter-events-adapter</module>
     </modules>
 
@@ -39,7 +39,7 @@
             </releases>
         </repository>
         <repository>
-	    <id>alfresco-internal-snapshots</id>
+            <id>alfresco-internal-snapshots</id>
             <url>https://artifacts.alfresco.com/nexus/content/repositories/snapshots</url>
         </repository>
     </repositories>
